Fix for Firefox
Firefox did not work as the the --moz selectors were not used. Adding the selectors helps, but there is enough other bits different between Firefox and Chrome that separate classes are required. New CSS tested on both Firefox and Chrome. Combining common styles would seem to be a better approach (less CSS code) however there's neither browser would recognize the second of the two combined styles. Changing the order of the `,` selector would fix one browser but break the other (and vice-versa).

I will submit a separate PR for the `switch-holder` css.